+++ application/models/AdsBatch.php
@@ -37,6 +37,7 @@ class AdsBatch extends MyAppModel
         $srch->joinTable(Language::DB_TBL, 'LEFT JOIN', self::DB_TBL_PREFIX . 'lang_id = ' . Language::DB_TBL_PREFIX . 'id', 'lang');
         $srch->joinTable(Countries::DB_TBL, 'LEFT JOIN', self::DB_TBL_PREFIX . 'target_country_id = ' . Countries::DB_TBL_PREFIX . 'id', 'c');
         if (true === $joinAdsProds) {
+            $now = FatDate::nowInTimezone(FatApp::getConfig('CONF_TIMEZONE'), 'Y-m-d');
             $srch->joinTable(self::DB_TBL_BATCH_PRODS, 'LEFT JOIN', self::DB_TBL_PREFIX . 'id = ' . self::DB_TBL_BATCH_PRODS_PREFIX . self::DB_TBL_PREFIX . 'id', 'abp');
             $srch->joinTable(SellerProduct::DB_TBL, 'INNER JOIN', 'selprod_id = abprod_selprod_id', 'sp');
             $srch->joinTable(SellerProduct::DB_TBL_LANG, 'INNER JOIN', 'selprod_id = selprodlang_selprod_id AND selprodlang_lang_id = ' . $langId, 'sp_l');
@@ -44,6 +45,21 @@ class AdsBatch extends MyAppModel
             $srch->joinTable(Product::DB_TBL_LANG, 'LEFT OUTER JOIN', 'p.product_id = p_l.productlang_product_id AND p_l.productlang_lang_id = ' . $langId, 'p_l');
             $srch->joinTable(Brand::DB_TBL, 'LEFT JOIN', 'b.brand_id = p.product_brand_id', 'b');
             $srch->joinTable(Brand::DB_TBL_LANG, 'LEFT JOIN', 'b_l.brandlang_brand_id = b.brand_id AND b_l.brandlang_lang_id = ' . $langId, 'b_l');
+
+            $srch->joinTable(
+                SellerProduct::DB_TBL_SELLER_PROD_SPCL_PRICE,
+                'LEFT OUTER JOIN',
+                'm.splprice_selprod_id = selprod_id AND \'' . $now . '\' BETWEEN m.splprice_start_date AND m.splprice_end_date',
+                'm'
+            );
+
+            $srch->joinTable(
+                SellerProduct::DB_TBL_SELLER_PROD_SPCL_PRICE,
+                'LEFT OUTER JOIN',
+                's.splprice_selprod_id = selprod_id AND s.splprice_price < m.splprice_price
+                 AND \'' . $now . '\' BETWEEN s.splprice_start_date AND s.splprice_end_date',
+                's'
+            );
         }
 
         return $srch;
@@ -118,7 +134,23 @@ class AdsBatch extends MyAppModel
         $srch->addCondition(AdsBatch::DB_TBL_PREFIX . 'user_id', '=', $userId);
         $srch->addMultipleFields(
             [
-                'selprod_id', 'selprod_title', 'selprod_stock', 'selprod_condition', 'selprod_price', 'selprod_available_from', 'product_id', 'product_description', 'product_upc', 'language_code', 'country_code', 'IFNULL(brand_name, brand_identifier) as brand_name', 'abprod_item_group_identifier', 'adsbatch_expired_on', 'abprod_cat_id'
+                'selprod_id',
+                'selprod_title',
+                'selprod_stock',
+                'selprod_condition',
+                'selprod_price',
+                'selprod_available_from',
+                'product_id',
+                'product_description',
+                'product_upc',
+                'language_code',
+                'country_code',
+                'IFNULL(brand_name, brand_identifier) as brand_name',
+                'abprod_item_group_identifier',
+                'adsbatch_expired_on',
+                'abprod_cat_id',
+                '(CASE WHEN m.splprice_selprod_id IS NULL THEN 0 ELSE 1 END) AS special_price_found',
+                'COALESCE(m.splprice_price, selprod_price) AS theprice'
             ]
         );       
         $productData = FatApp::getDb()->fetchAll($srch->getResultSet());