+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 - Bug-070110 - User dashboard area breadcrumb labels shall be manageable from admin end
 - Bug-070125 - Testimonial image is not rendering incase of other language
 - Bug-070173 - shopify data migration not worrking as tag lang table not exisit in current release
+- Bug-070827 - Adding special price to batch products in google shopping feed.
 # RV-10.0.0.20221123
 
 ## Bug Fixes: